Reconstruction after the review
