+++ app/src/main/java/com/iyanuoluwa/imovie/ui/common/NaughtyPager.kt
@@ -1,4 +1,4 @@
-package com.iyanuoluwa.imovie.util.custom
+package com.iyanuoluwa.imovie.ui.common
 
 import android.content.Context
 import android.util.AttributeSet
