+++ app/src/main/java/com/iyanuoluwa/imovie/ui/main/Popular.kt
@@ -1,4 +1,4 @@
-package com.iyanuoluwa.imovie
+package com.iyanuoluwa.imovie.ui.main
 
 import android.os.Bundle
 import android.util.Log
@@ -11,9 +11,10 @@ import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.GridLayoutManager
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import com.iyanuoluwa.imovie.ui.adapter.MovieAdapter
-import com.iyanuoluwa.imovie.data.model.buffer.api.MovieJson
-import com.iyanuoluwa.imovie.data.remote.request.ApiTopRated
+import com.iyanuoluwa.imovie.R
+import com.iyanuoluwa.imovie.data.model.MovieJson
+import com.iyanuoluwa.imovie.data.remote.ApiPopular
+import com.iyanuoluwa.imovie.ui.details.MovieAdapter
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
@@ -21,9 +22,9 @@ import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
 
 
-class TopRated : Fragment() {
+class Popular : Fragment() {
 
-    private var topRatedRecyclerView: RecyclerView? =null
+    private var popularRecyclerView: RecyclerView? =null
     private var gridLayoutManager: GridLayoutManager? = null
     private var titleList = mutableListOf<String>()
     private var imageList = mutableListOf<String>()
@@ -38,15 +39,15 @@ class TopRated : Fragment() {
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
                               savedInstanceState: Bundle?): View? {
         // Inflate the layout for this fragment
-        val view = inflater.inflate(R.layout.fragment_top_rated, container, false)
-        topRatedRecyclerView = view.findViewById(R.id.top_rated_recycler_view)
+        val view = inflater.inflate(R.layout.fragment_popular, container, false)
+        popularRecyclerView = view.findViewById(R.id.popular_recycler_view)
         gridLayoutManager = GridLayoutManager(context, 3, LinearLayoutManager.VERTICAL, false)
-        nestedScrollView = view.findViewById(R.id.scroll_view_topRated)
-        progressBar = view.findViewById(R.id.progress_bar_topRated)
+        nestedScrollView = view.findViewById(R.id.scroll_view_popular)
+        progressBar = view.findViewById(R.id.progress_bar_popular)
         return view
     }
 
-    private fun addToList(title: String, image: String, plot: String, id: Int) {
+    private fun addToList(title : String, image: String, plot: String, id: Int) {
         titleList.add(title)
         imageList.add(image)
         plotList.add(plot)
@@ -54,39 +55,39 @@ class TopRated : Fragment() {
     }
 
     private fun setUpRecyclerView() {
-        topRatedRecyclerView?.layoutManager = gridLayoutManager
-        topRatedRecyclerView?.adapter = context?.let { MovieAdapter(it, titleList, imageList, plotList, ids) }
+        popularRecyclerView?.layoutManager = gridLayoutManager
+        popularRecyclerView?.adapter = context?.let { MovieAdapter(it, titleList, imageList, plotList, ids) }
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        getTopRatedMovies(page, limit)
+        getPopularMovies(page, limit)
         nestedScrollView?.setOnScrollChangeListener(object : NestedScrollView.OnScrollChangeListener {
             override fun onScrollChange(v: NestedScrollView?, scrollX: Int, scrollY: Int, oldScrollX: Int, oldScrollY: Int) {
                 if (scrollY == (v?.getChildAt(0)?.measuredHeight)?.minus(v?.measuredHeight!!)) {
                     page++
                     progressBar?.visibility = View.VISIBLE
-                    getTopRatedMovies(page, limit)
+                    getPopularMovies(page, limit)
                 }
             }
         })
     }
 
-    private fun getTopRatedMovies(page: Int, limit: Int) {
-        val apiTopRated = Retrofit.Builder()
+    private fun getPopularMovies(page: Int, limit: Int) {
+        val apiPopular = Retrofit.Builder()
                 .baseUrl(BASE_URL)
                 .addConverterFactory(GsonConverterFactory.create())
                 .build()
-                .create(ApiTopRated::class.java)
+                .create(ApiPopular::class.java)
 
-        val data = apiTopRated.getTopRatedMovies(page, limit)
+        val data = apiPopular.getPopularMovies(page, limit)
         data.enqueue(object : Callback<MovieJson?> {
             override fun onResponse(call: Call<MovieJson?>, response: Response<MovieJson?>) {
                 if (response.isSuccessful && response.body() != null) {
                     progressBar?.visibility = View.GONE
                     val responseBody = response.body()!!
                     for (movies in responseBody.results) {
-                        Log.i("TopRatedFragment", "Result = $movies")
+                        Log.i("PopularFragment", "Result = $movies")
                         addToList(movies.title, "https://image.tmdb.org/t/p/w500${movies.posterPath}", movies.overview, movies.id)
                     }
                     setUpRecyclerView()
@@ -94,7 +95,7 @@ class TopRated : Fragment() {
             }
 
             override fun onFailure(call: Call<MovieJson?>, t: Throwable) {
-                Log.e("TopRatedFragment", t.toString())
+                Log.e("PopularFragment", t.toString())
             }
         })
     }
