+++ app/src/main/res/layout/fragment_popular.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".Popular">
+    tools:context=".ui.main.Popular">
 
     <TextView
         android:id="@+id/popular"
