+++ app/src/main/res/layout/fragment_top_rated.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".TopRated">
+    tools:context=".ui.main.TopRated">
 
     <TextView
         android:id="@+id/top_rated"
