+++ app/src/main/res/layout/fragment_up_coming.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:orientation="vertical"
     android:layout_height="match_parent"
-    tools:context=".UpComing">
+    tools:context=".ui.main.UpComing">
 
     <TextView
         android:id="@+id/up_coming"
